# tools/admin-api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Open5GS Admin API
  version: "0.1.0"
  description: |
    Append-only admin API for dynamic MME/SMF/UPF configuration.

    Semantics:
      * **POST** adds a resource if its natural key is not already present.
        Re-posting an identical resource is a no-op (returns 200 with the
        existing doc, does **not** bump the global revision).
      * **DELETE** removes the configuration row. It does NOT touch live UEs,
        eNBs, or gNBs. Cleanup of in-flight sessions is out of scope.
      * A monotonic global revision is bumped on every successful mutating
        call. NFs poll the admin API (or tail Mongo directly) and apply the
        delta on the next tick. Clients can watch `/apply-status` to see
        which NFs have converged.
  license:
    name: AGPL-3.0-only
servers:
  - url: http://localhost:9998
    description: local dev
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    UpfPeer:
      type: object
      required: [host, port]
      properties:
        id:        { type: string, readOnly: true }
        host:      { type: string }
        port:      { type: integer, default: 8805 }
        dnns:
          type: array
          items: { type: string }
        label:     { type: string, nullable: true }
        revision:  { type: integer, format: int64, readOnly: true }
    UpfPeerCreate:
      type: object
      required: [host]
      properties:
        host:  { type: string }
        port:  { type: integer, default: 8805 }
        dnns:
          type: array
          items: { type: string }
        label: { type: string, nullable: true }
...
paths:
...
  /api/v1/upf-peers:
    get:
      summary: List UPF peers known to SMF
      tags: [upf-peers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UpfPeer' }
    post:
      summary: Add a UPF peer
      tags: [upf-peers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpfPeerCreate' }
      responses:
        "201": { description: Created }
        "200": { description: Already existed }
        "400": { description: Validation error }
        "409": { description: Capacity reached }
  /api/v1/upf-peers/{host}/{port}:
    delete:
      summary: Delete a UPF peer
      tags: [upf-peers]
      parameters:
        - name: host
          in: path
          required: true
          schema: { type: string }
        - name: port
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204": { description: Deleted }
        "404": { description: Not found }
...